Fix logged pre-launch issues
This adds a new field to indicate that a suggestion has been reviewed, that fixes #2, also changes classifier form to require name and closes #3. Modifies admin proposal workflow to prevent the editing of a proposal after it has been voted on and closes #4, alongside changing the proposal voting workflow to prevent changes to submitted votes and resolves #5.